Fix method arg count store for invokeBasic
The VM uses the value stored in VMThread.tempSlot to determine
the location of the receiver object in the stack for the
signature-polymorphic calls to invokeBasic, accessing the entry
at _sp[VMThread.tempSlot].
However, the JIT was storing the number of args in the
invokeBasic call into VMThread.tempSlot, which assumes each arg
for the invokeBasic call occupies one stack slot. This was not
the case for data types that occupied more than one stack slot,
for example, doubles.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>