Check the InnerClass attribute of the enclosing class
The change is to check to see whether an valid entry for
the specified class exists in the InnerClass attribute
of its enclosing class given the InnerClass attribute
of both the inner class and its enclosing class must
remain consistent with each other.

The idea is to add all inner class entries in the InnerClass
attribute to the enclosing class, in which case those
inner class which are defined inside a method or defined
inside a nested class should be skipped in getDeclaredClassesImpl
and isClassADeclaredClass.

Signed-off-by: Cheng Jin <jincheng@ca.ibm.com>