Implement thunk-gen and getArgPointer for AIX64
Since AIX ABI accommodates the back-up stack area for all parameters,
we always push them back to the caller frame. In case, a struct is
returned, we need to take the hidden parameter into account. As a
result, we need to buy a stack frame for java to return and distribute
the result back.

Straight-forward calculation for each argument's offset.

Signed-off-by: Julian <zlwang@ca.ibm.com>